% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_autodiff.R
\name{available_sensitivity}
\alias{available_sensitivity}
\title{Check what sensitivities are available}
\usage{
available_sensitivity(res, display = T)
}
\arguments{
\item{res}{Output from *_AD function.}

\item{display}{T or F; if T, print out summary.}
}
\description{
Check what sensitivities are available
}
