% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesSense.R
\name{BayesSense}
\alias{BayesSense}
\title{Sensitivity analysis for Bayesian Inference}
\description{
This package performs a comprehensive local sensitivity analysis of MCMC
output with respect to all input parameters (prior hyper-parameters and starting values) to assess prior robustness and algorithm convergence in
models estimated via Gibbs samplers.
It extends the usual Automatic Differentiation to allow for differentiation of
matrix decomposition and random samples, and it exploits matrix and vector calculus results
to compute large sets of Jacobian matrices that contain the first order
derivatives of all MCMC draws with respect to all input parameters.
A range of sensitivity measures for posterior statistics can be computed,
such as prior robustness of posterior means and prediction intervals,
as well as algorithm convergence measures bases on starting value sensitivities.
}
\details{
Package 'BayesSense'
}
\author{
Liana Jacobi [aut, cph], Dan Zhu [aut, cph], Chun Fung Kwok [aut, cre], Kai-Yang Goh [ctb]
}
