% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_posterior_stat.R
\name{burn_ins}
\alias{burn_ins}
\title{Remove the burn-in samples}
\usage{
burn_ins(res, num_burns)
}
\arguments{
\item{res}{The Jacobian array.}

\item{num_burns}{Positive integer; the number of samples to burn.}
}
\description{
Remove the burn-in samples
}
