% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_posterior_stat.R
\name{cumulative_stat}
\alias{cumulative_stat}
\title{Cumulative summary statistics}
\usage{
cumulative_stat(x, stat_fun, ...)
}
\arguments{
\item{x}{numeric vector.}

\item{stat_fun}{function; the summary statistics function.}

\item{...}{Extra parameters to be passed to stat_fun.}
}
\description{
Cumulative summary statistics
}
