% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VAR_forecast.R
\name{fitted_VAR}
\alias{fitted_VAR}
\title{Generate fitted values given a VAR model}
\usage{
fitted_VAR(data0, b_0, B)
}
\arguments{
\item{data0}{Matrix; each row is one observation, each column is one measurement / predictor.}

\item{b_0}{numeric vector; the intercept.}

\item{B}{list of numeric matrices; the regression coefficients.}
}
\description{
Generate fitted values given a VAR model
}
