% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian_autodiff.R
\name{gaussian_AD}
\alias{gaussian_AD}
\title{Sensitivity analysis for normal regression model with multivariate normal
prior for the mean and inverse gamma for the variance.}
\usage{
gaussian_AD(X, y, b_0, B_0, alpha_0, delta_0, init_sigma,
  num_steps = 1000)
}
\arguments{
\item{X}{A numeric matrix; the covariates.}

\item{y}{A numeric vector; the response variable.}

\item{b_0}{A numeric vector; the mean for the multivariate normal prior.}

\item{B_0}{A numeric matrix; the covariance for the multivariate normal prior.}

\item{alpha_0}{A positive number; the shape parameter of the inverse-gamma prior.}

\item{delta_0}{A positive number; the rate parameter of the inverse-gamma prior.}

\item{init_sigma}{(Optional) A numeric matrix}

\item{num_steps}{Integer; number of MCMC steps.}
}
\description{
Sensitivity analysis for normal regression model with multivariate normal
prior for the mean and inverse gamma for the variance.
}
\examples{
\dontrun{
n <- 1000
p <- 5
data0 <- gaussian_data(n, p, intercept = TRUE)
res <- gaussian_AD(
  X = data0$X, y = data0$y,
  b_0 = rnorm(p+1), B_0 = diag(p+1),  # add one for the intercept
  alpha_0 = 13, delta_0 = 8
)
}
}
