% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{gaussian_Gibbs}
\alias{gaussian_Gibbs}
\title{Model inference for normal regression model with multivariate normal
prior for the mean and inverse gamma for the variance.}
\usage{
gaussian_Gibbs(X, y, b_0, B_0, alpha_0, delta_0, beta, sigma,
  num_steps = 10000, burn_ins = 1000)
}
\arguments{
\item{X}{A numeric matrix; the covariates.}

\item{y}{A numeric vector; the response variable.}

\item{b_0}{A numeric vector; the mean for the multivariate normal prior.}

\item{B_0}{A numeric matrix; the covariance for the multivariate normal prior.}

\item{alpha_0}{A positive number; the shape parameter of the inverse-gamma prior.}

\item{delta_0}{A positive number; the rate parameter of the inverse-gamma prior.}

\item{beta}{(Optional) A numeric vector}

\item{sigma}{(Optional) A numeric matrix}

\item{num_steps}{Integer; number of MCMC steps.}

\item{burn_ins}{Integer; number of burn-ins.}
}
\description{
Model inference for normal regression model with multivariate normal
prior for the mean and inverse gamma for the variance.
}
\examples{
\dontrun{
n <- 100
p <- 10
data0 <- gaussian_data(n, p)
res <- gaussian_Gibbs(data0$X, data0$y,
  b_0 = rnorm(p), B_0 = pdmatrix(p)$Sigma, alpha_0 = 13, delta_0 = 8,
)
}
}
