% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_autodiff.R
\name{get_sensitivity}
\alias{get_sensitivity}
\title{Retrieve a sensitivitiy from autodiff output}
\usage{
get_sensitivity(res, numerator, denominator, reshape = T)
}
\arguments{
\item{res}{Output from *_AD function.}

\item{numerator}{Character string; the numerator from `available_sensitivity`.}

\item{denominator}{Character string; the denominator from `available_sensitivity`.}

\item{reshape}{T or F; if T, reshape the result into an array.}
}
\description{
Retrieve a sensitivitiy from autodiff output
}
