% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{init_differential}
\alias{init_differential}
\title{Initialise internal input for Autodiff function}
\usage{
init_differential(len0, vec0, names0)
}
\arguments{
\item{len0}{the dimension of the numerator}

\item{vec0}{the vector of dimensions of the denominator (parameters)}

\item{names0}{the vector of names of the denominator (parameters)}
}
\description{
This function initialises list of zero matrices.
}
