% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{IPA}
\alias{IPA}
\title{Infinitesimal Perturbation Analysis}
\usage{
IPA(model0, X, y, param, change, pos = 1, h, stat_fun = mean)
}
\arguments{
\item{model0}{Function; any one of *-Gibbs functions.}

\item{X}{Numeric matrix; the covariates.}

\item{y}{Numeric vector; the response variable.}

\item{param}{List; a list of named parameters.}

\item{change}{String; name of the parameter to change.}

\item{pos}{Integer; if the parameter is a vector, then provide the position
of the specific parameter.}

\item{h}{Number; the step size for the sensitivitiy analysis.}

\item{stat_fun}{Function; the summary function of the posterior distribution.}
}
\description{
Infinitesimal Perturbation Analysis
}
\examples{
\dontrun{
b_0 <- rnorm(p)
B_0 <- pdmatrix(p)$Sigma
alpha_0 <- 13
delta_0 <- 8
# Analyse the sensitivity of posterior mean w.r.t. parameter 'alpha_0'.
h <- 0.01
stat_fun <- mean
IPA(
  gaussian_Gibbs, data0$X, data0$y,
  param = list(b_0 = b_0, B_0 = B_0, alpha_0 = alpha_0, delta_0 = delta_0),
  change = "alpha_0", h = h, stat_fun = stat_fun
)
}
}
