% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_posterior_stat.R
\name{jacobian}
\alias{jacobian}
\title{Jacobian matrix}
\usage{
jacobian(x, dim0)
}
\arguments{
\item{x}{the sensitivity of the posterior samples}

\item{dim0}{the dimension of the Jacobian matrix}
}
\description{
Jacobian matrix
}
