% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_numerical_differentiation.R
\name{num_diff}
\alias{num_diff}
\title{Numerical differentiation}
\usage{
num_diff(f, fixed_param, vary_param, sum_stat, delta, seed)
}
\arguments{
\item{f}{a function to be differentiated numerically.}

\item{fixed_param}{a list of fixed parameters to be called upon f.}

\item{vary_param}{a list of parameters of which the deriatives are seeked.}

\item{sum_stat}{a function to compute summary statistics.}

\item{delta}{a positive number, the amount of perturbation.}
}
\description{
Numerical differentiation
}
