% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_autoregressive.R
\name{param_posterior_mean}
\alias{param_posterior_mean}
\title{Compute the posterior mean of the parameters}
\usage{
param_posterior_mean(model0, n, p)
}
\arguments{
\item{model0}{Output from 'VAR_Gibbs'.}

\item{n}{integer; dimension of the data.}

\item{p}{integer; lag of the time series model.}
}
\description{
Compute the posterior mean of the parameters
}
