% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{pdmatrix}
\alias{pdmatrix}
\title{This is a function to generate covariance matrix.}
\usage{
pdmatrix(...)
}
\arguments{
\item{...}{Parameters to be passed to 'genPositiveDefMat'.
See '?clusterGeneration::genPositiveDefMat' for documentation.}
}
\description{
This is a function to generate covariance matrix.
}
\details{
Alias for 'clusterGeneration::genPositiveDefMat'
}
\examples{
# Create a random 4 x 4 covariance matrix
cov_mat <- pdmatrix(dim = 4)$Sigma
print(cov_mat)
}
