% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_plot.R
\name{plot_mts}
\alias{plot_mts}
\title{Plotting function for multivariate time series.}
\usage{
plot_mts(data0, x_axis, ...)
}
\arguments{
\item{data0}{Numeric matrix; each column is one time series.}

\item{x_axis}{Vector (optional); the break of the axis.}

\item{...}{Other parameters to be passed to the plot functions,
e.g. 'xlab', 'ylab' and 'main'.}
}
\description{
Plotting function for multivariate time series.
}
