% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_posterior}
\alias{plot_posterior}
\title{Plot the posterior distribution of the parameters}
\usage{
plot_posterior(res)
}
\arguments{
\item{res}{Output from any *-Gibbs functions.}
}
\description{
Plot the posterior distribution of the parameters
}
