% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{s_eig_matrix}
\alias{s_eig_matrix}
\title{This is a function to generate matrix with eigenvalue less than 1 based on
some heuristic.}
\usage{
s_eig_matrix(dim)
}
\arguments{
\item{dim}{integer; dimension of the matrix.}
}
\description{
This is a function to generate matrix with eigenvalue less than 1 based on
some heuristic.
}
\examples{
# Create a random 4 x 4 matrix with eigenvalue less than 1.
mat <- s_eig_matrix(4)
print(mat)
eigen(mat)$value
}
