% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VAR.R
\name{simulate_VAR}
\alias{simulate_VAR}
\title{Simulate data from a Vector AutoRegressive (VAR) model}
\usage{
simulate_VAR(m, n, p, init_y_t, b_0, B, Sigma)
}
\arguments{
\item{m}{integer; number of datapoints.}

\item{n}{integer; dimension of the observation.}

\item{p}{integer; the time-series lag.}

\item{init_y_t}{numeric matrix (optional); the initial state of the time series.
The matrix must have as many columns as the lag, and the first
column should be the state at the first time point.}

\item{b_0}{numeric vector (optional); the intercept.}

\item{B}{list of numeric matrices (optional); the regression coefficients.}

\item{Sigma}{numeric matrix (optional); the covariance matrix of the error.}
}
\description{
Simulate data from a Vector AutoRegressive (VAR) model
}
\examples{
simulate_VAR(30, 3, 2)
}
