% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/student_t_autodiff.R
\name{student_t_AD}
\alias{student_t_AD}
\title{Sensitivity analysis for student-t regression model with multivariate normal
prior for the mean and inverse gamma for the variance.}
\usage{
student_t_AD(X, y, b_0, B_0, alpha_0, delta_0, nu, init_beta, init_sigma,
  num_steps = 10000)
}
\arguments{
\item{X}{A numeric matrix; the covariates.}

\item{y}{A numeric vector; the response variable.}

\item{b_0}{A numeric vector; the mean for the multivariate normal prior.}

\item{B_0}{A numeric matrix; the covariance for the multivariate normal prior.}

\item{alpha_0}{A positive number; the shape parameter of the inverse-gamma prior.}

\item{delta_0}{A positive number; the rate parameter of the inverse-gamma prior.}

\item{nu}{A positive number; the shape and rate parameters for the gamma prior.}

\item{init_beta}{(Optional) A numeric vector; the starting value of the beta parameter.}

\item{init_sigma}{(Optional) A positive number; the starting value of the sigma parameter.}

\item{num_steps}{Integer; number of MCMC steps.}

\item{burn_ins}{Integer; number of burn-ins.}
}
\description{
Sensitivity analysis for student-t regression model with multivariate normal
prior for the mean and inverse gamma for the variance.
}
\examples{
\dontrun{
n <- 50
p <- 3
data0 <- student_t_data(n, p, intercept = TRUE)
res <- student_t_AD(data0$X, data0$y,
  b_0 = rnorm(p+1), B_0 = diag(p+1),  # add one for intercept
  alpha_0 = 13, delta_0 = 8, nu = 5
)
}
}
