% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{student_t_data}
\alias{student_t_data}
\title{Simulate data from a linear regression model with student-t noise}
\usage{
student_t_data(n, p, beta, df, intercept = TRUE)
}
\arguments{
\item{n}{integer; number of datapoints.}

\item{p}{integer; number of measurements.}

\item{beta}{(optional) numeric vector; the regression coefficients.
It should be of length (p+1) if 'intercept = T', and the intercept
should be the first element.}

\item{df}{positive number; degrees of freedom, Inf is allowed.}

\item{intercept}{TRUE or FALSE; whether to include an intercept / bias term.}
}
\value{
A list of the data X, y, and the parameters beta, df.
}
\description{
Simulate data from a linear regression model with student-t noise
}
\examples{
data0 <- student_t_data(n = 100, p = 10)
data1 <- student_t_data(n = 100, p = 3, beta = c(0.2, -0.3, 0.4), df = 2, intercept = FALSE)
}
