% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VAR_forecast.R
\name{summarise_forecast}
\alias{summarise_forecast}
\title{Summarise forecast posterior density}
\usage{
summarise_forecast(arr_forecast, method, ...)
}
\arguments{
\item{arr_forecast}{Output from VAR_posterior forecast}

\item{method}{function; summary statistical function, e.g. mean, quantile, etc.}

\item{...}{Other parameters to be passed to the 'method'.}
}
\description{
Summarise forecast posterior density
}
