% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_posterior_stat.R
\name{thinning}
\alias{thinning}
\title{Thinning the samples}
\usage{
thinning(vec0, every_n)
}
\arguments{
\item{vec0}{A numeric vector; the samples.}

\item{every_n}{Positive integer; take one sample every_n sample.}
}
\description{
Thinning the samples
}
