% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{tidy_list}
\alias{tidy_list}
\title{Tidy internal output for Autodiff function}
\usage{
tidy_list(dlist0)
}
\arguments{
\item{dlist0}{expects (a list of) N lists of matrices named by vec0.}
}
\value{
a named list of (N-row) matrices.
}
\description{
Tidy internal output for Autodiff function
}
