% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VAR_util.R
\name{train_data}
\alias{train_data}
\title{#' Convert the data into a format ready for modelling
#' @param data0 Matrix; the number of columns is equal to the number of time points,
#' the number of rows is equal to the dimension of the observation. One can get
#' such data from the first component of the output from 'simulate_VAR'.
#' @param p integer; time series lag.
#' @export
VAR_model_matrix <- function(data0, p) {
  eff_T <- ncol(data0) - p
  n <- nrow(data0)
  I_n <- diag(n)
  purrr::map2(
    .x = seq(eff_T), .y = p + seq(eff_T) - 1,
    .f = ~ I_n %x% t( c(1, matrixcalc::vec(data0[, .y:.x, drop = F])) )
  ) %>%
    do.call(rbind, .)
}}
\usage{
train_data(Y, lag)
}
\description{
#' Convert the data into a format ready for modelling
#' @param data0 Matrix; the number of columns is equal to the number of time points,
#' the number of rows is equal to the dimension of the observation. One can get
#' such data from the first component of the output from 'simulate_VAR'.
#' @param p integer; time series lag.
#' @export
VAR_model_matrix <- function(data0, p) {
  eff_T <- ncol(data0) - p
  n <- nrow(data0)
  I_n <- diag(n)
  purrr::map2(
    .x = seq(eff_T), .y = p + seq(eff_T) - 1,
    .f = ~ I_n %x% t( c(1, matrixcalc::vec(data0[, .y:.x, drop = F])) )
  ) %>%
    do.call(rbind, .)
}
}
\keyword{internal}
