% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VAR_autodiff.R
\name{VAR_AD}
\alias{VAR_AD}
\title{Sensitivity analysis for Vector-Auto-Regressive model with normal-Wishart priors.}
\usage{
VAR_AD(data0, lag, b_0, B_0, v_0, S_0, init_Sigma, num_steps = 3000,
  burn_ins = 1000)
}
\arguments{
\item{data0}{Matrix; each row is one observation, each column is one measurement / predictor.}

\item{lag}{Integer; the lag of the time series model.}

\item{b_0}{A numeric vector; the mean for the multivariate normal prior.}

\item{B_0}{A numeric matrix; the covariance for the multivariate normal prior.}

\item{v_0}{scalar; degree of freedom for the inverse-Wishart prior.}

\item{S_0}{matrix; the scale matrix for the inverse-Wishart prior.}

\item{init_Sigma}{(Optional) matrix; the starting value of the noise inverse-covariance.}

\item{num_steps}{integer; number of MCMC steps.}

\item{burn_ins}{integer; number of burn-ins.}
}
\description{
Sensitivity analysis for Vector-Auto-Regressive model with normal-Wishart priors.
}
