% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VAR.R
\name{VAR_Gibbs}
\alias{VAR_Gibbs}
\title{Model inference for Bayesian Vector-AutoRegressive (VAR) model with
independent normal-inverse-Wishart priors.}
\usage{
VAR_Gibbs(data0, lag, b_0, B_0, v_0, S_0, init_Sigma, num_steps = 3000,
  burn_ins = 1000)
}
\arguments{
\item{data0}{Matrix; each row is one observation, each column is one measurement / predictor.}

\item{lag}{Integer; the lag of the time series model.}

\item{b_0}{A numeric vector; the mean for the multivariate normal prior.}

\item{B_0}{A numeric matrix; the covariance for the multivariate normal prior.}

\item{v_0}{scalar; degree of freedom for the inverse-Wishart prior.}

\item{S_0}{matrix; the scale matrix for the inverse-Wishart prior.}

\item{init_Sigma}{(Optional) matrix; the starting value of the noise inverse-covariance.}

\item{num_steps}{integer; number of MCMC steps.}

\item{burn_ins}{integer; number of burn-ins.}
}
\value{
A list of two matrices. The first matrix contains the posterior samples
of beta and the second matrix contains the posterior samples of Sigma. In both
matrices, each row corresponding to 1 sample of the vectorised parameter.
}
\description{
Model inference for Bayesian Vector-AutoRegressive (VAR) model with
independent normal-inverse-Wishart priors.
}
\examples{
sim <- simulate_VAR(m = 1000, 3, 2) # 1000 pts of 3-dimensional lag-2 time series
data0 <- sim$data

dim_data <- 3
lag <- 2
d <- dim_data + dim_data ^2 * lag  # dimension of the vectorised parameter

res <- VAR_Gibbs(
  data0, lag,
  b_0 = rnorm(d), B_0 = pdmatrix(d)$Sigma,
  v_0 = 3, S_0 = pdmatrix(dim_data)$Sigma,
  init_Sigma = diag(dim_data)
)
}
