% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_autoregressive.R
\name{VAR_model_matrix}
\alias{VAR_model_matrix}
\title{Convert the data into a format ready for modelling}
\usage{
VAR_model_matrix(data0, p)
}
\arguments{
\item{data0}{Matrix; the number of columns is equal to the number of time points,
the number of rows is equal to the dimension of the observation. One can get
such data from the first component of the output from 'simulate_VAR'.}

\item{p}{integer; time series lag.}
}
\description{
Convert the data into a format ready for modelling
}
