% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VAR_util.R
\name{VAR_model_to_vec}
\alias{VAR_model_to_vec}
\title{Vectorise all the coefficients appeared in the model}
\usage{
VAR_model_to_vec(model0)
}
\arguments{
\item{model0}{A VAR model; the second component of the output from 'simulate_VAR'.}
}
\description{
Vectorise all the coefficients appeared in the model
}
