% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VAR_summary.R
\name{VAR_posterior_forecast}
\alias{VAR_posterior_forecast}
\title{Compute the posterior forecast}
\usage{
VAR_posterior_forecast(model0, data0, n_step_ahead)
}
\arguments{
\item{model0}{Output from 'VAR_Gibbs'.}

\item{data0}{Matrix; each row is one observation, each column is one measurement / predictor.}

\item{n_step_ahead}{integer; number of future forecast to generate.}
}
\value{
A array ARR where ARR[, , k] contains the forecast of the posterior sample at
iteration k. ARR[, , k] is of dimension D x n_step_ahead, where D is the dimension
of the multivariate time series.
}
\description{
This function takes each posterior sample and perform a n-step
forecast on the given time series.
}
