% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VAR_util.R
\name{VAR_posterior_stat}
\alias{VAR_posterior_stat}
\title{Compute the posterior statistics of the parameters}
\usage{
VAR_posterior_stat(model0, n, p, fun = mean)
}
\arguments{
\item{model0}{Output from 'VAR_Gibbs'.}

\item{n}{integer; dimension of the data.}

\item{p}{integer; lag of the time series model.}

\item{fun}{function; statistical function, default set to be 'mean'.}
}
\description{
Compute the posterior statistics of the parameters
}
