% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VAR_summary.R
\name{VAR_summarise_forecast}
\alias{VAR_summarise_forecast}
\title{Summarise forecast posterior density}
\usage{
VAR_summarise_forecast(arr_forecast, method, ...)
}
\arguments{
\item{arr_forecast}{Output from VAR_posterior forecast}

\item{method}{function; summary statistical function, e.g. mean, quantile, etc.}

\item{...}{Other parameters to be passed to the 'method'.}
}
\value{
A list of summary statistics. The list should have length D, where D is
the dimension of the multivariate time series. Each component contains a matrix
with dimension K x n_step_ahead, where K is the dimension of the summary statistics
and n_step_ahead is the parameter set in "VAR_posterior_forecast".
}
\description{
Summarise forecast posterior density
}
