% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_autoregressive.R
\name{VAR_vec_to_model_coeff}
\alias{VAR_vec_to_model_coeff}
\title{Convert vector back to model coefficients}
\usage{
VAR_vec_to_model_coeff(vec0, n)
}
\arguments{
\item{vec0}{Numeric vector; vectorised model coefficients.}

\item{n}{integer; dimension of the data.}
}
\description{
Convert vector back to model coefficients
}
