% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wishart_autodiff.R
\name{wishart_AD}
\alias{wishart_AD}
\title{Sensitivity analysis for Wishart regression model with multivariate normal
prior for the means and inverse gamma for the variance.}
\usage{
wishart_AD(Xy, y, b_0, B_0, Xs, s, g_0, G_0, v_0, R_0, init_gamma,
  init_Sigma, num_steps = 10000)
}
\arguments{
\item{Xy}{A numeric matrix; the first set of covariates.}

\item{y}{A numeric vector; the response variable for the first set of covariates.}

\item{b_0}{A numeric vector; the mean for the first multivariate normal prior.}

\item{B_0}{A numeric matrix; the covariance for the first multivariate normal prior.}

\item{Xs}{A numeric matrix; the second set of covariates.}

\item{s}{A numeric vector; the response variable for the second set of covariates.}

\item{g_0}{A numeric vector; the mean for the second multivariate normal prior.}

\item{G_0}{A numeric matrix; the covariance for the second multivariate normal prior.}

\item{v_0}{A positive number; the degrees of freedom of the Wishart prior.}

\item{R_0}{A 2x2 symmetric matrix; the scale matrix of the Wishart prior.}

\item{init_gamma}{(Optional) A numeric vector; the starting value of the gamma parameter.}

\item{init_Sigma}{(Optional) A 2x2 symmetric matrix; the starting value of the sigma parameter.}

\item{num_steps}{Integer; number of MCMC steps.}
}
\description{
Sensitivity analysis for Wishart regression model with multivariate normal
prior for the means and inverse gamma for the variance.
}
\examples{
\dontrun{
n <- 1000
p <- 2
k <- 3
data0 <- wishart_data(n, p, k,
                      Sigma = matrix(c(1, 0.2, 0.2, 1), 2, 2),
                      intercept_1 = TRUE, intercept_2 = TRUE)
res <- wishart_AD(Xy = data0$Xy, y = data0$y,
                  b_0 = numeric(p + 2), B_0 = diag(p + 2),
                  Xs = data0$Xs, s = data0$s,
                  g_0 = numeric(k + 1), G_0 = diag(k + 1),
                  v_0 = 5, R_0 = diag(2),
                  num_steps = 1e3)
}
}
