% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wishart_simulation.R
\name{wishart_data}
\alias{wishart_data}
\title{Simulate data from a 2-equation linear regression model with jointly Gaussian noise}
\usage{
wishart_data(n, p, k, beta, gamma, sigma, intercept_1 = TRUE,
  intercept_2 = TRUE)
}
\arguments{
\item{n}{integer; number of datapoints.}

\item{p}{integer; number of measurements of first set of covariates.}

\item{k}{integer; number of measurements of second set of covariates.}

\item{beta}{(optional) numeric vector; the first set of regression coefficients.
It should be of length (p+1) if 'intercept_1 = T', and the intercept
should be the first element.}

\item{gamma}{(optional) numeric vector; the second set of regression coefficients.
It shoud be of length (k+2) if 'intercept_2 = T'. The intercept should be
the first element. The endogeneity factor should be the second element.}

\item{sigma}{(optional) 2x2 numeric matrix; the covariance matrix of the distribution.}

\item{intercept_1}{TRUE or FALSE; whether to include an intercept / bias term.}

\item{intercept_2}{TRUE or FALSE; whether to include an intercept / bias term.}
}
\value{
A list of the data X, y, and the parameters beta, df.
}
\description{
Simulate data from a 2-equation linear regression model with jointly Gaussian noise
}
\examples{
data0 <- wishart_data(n = 100, p = 10, k = 3)
data1 <- wishart_data(n = 100, p = 2, k = 3,
  beta = c(0.2, -0.3, 0.4), gamma = c(0.2, 0.6, -0.2),
  sigma = matrix(c(1, 0.2, 0.2, 1), 2, 2), intercept_1 = FALSE, intercept_2 = FALSE)
}
