% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoBayes.R
\name{BayesSense}
\alias{BayesSense}
\title{Sensitivity analysis for Bayesian Inference}
\description{
This package computes local derivatives of Bayesian outputs, e.g.
MCMC posterior estimates and performance measures, via Automatic differentiation.
}
\details{
Package 'BayesSense'
}
\author{
Liana Jacobi <ljacobi@unimelb.edu.au>,
Dan Zhu <dan.zhu@monash.edu>,
Chun Fung Kwok <kcf.jackson@gmail.com>,
Kai-Yang Goh <kaiyanggoh@hotmail.com>
}
