% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VAR_util.R
\name{fitted_VAR}
\alias{fitted_VAR}
\title{Generate fitted values given a VAR model}
\usage{
fitted_VAR(data0, b_0, B)
}
\arguments{
\item{data0}{numeric matrix; the multivariate time series data.
The matrix must have at least as many columns as the lag, and the first
column should be the state at the first time point.}

\item{b_0}{numeric vector; the intercept.}

\item{B}{list of numeric matrices; the regression coefficients.}
}
\description{
Generate fitted values given a VAR model
}
