% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian_inference.R
\name{gaussian_Gibbs}
\alias{gaussian_Gibbs}
\title{Model inference for normal regression model with multivariate normal
prior for the mean and inverse gamma for the variance.}
\usage{
gaussian_Gibbs(X, y, b_0, B_0, alpha_0, delta_0, init_sigma,
  num_steps = 10000)
}
\arguments{
\item{X}{A numeric matrix; the covariates.}

\item{y}{A numeric vector; the response variable.}

\item{b_0}{A numeric vector; the mean for the multivariate normal prior.}

\item{B_0}{A numeric matrix; the covariance for the multivariate normal prior.}

\item{alpha_0}{A positive number; the shape parameter of the inverse-gamma prior.}

\item{delta_0}{A positive number; the rate parameter of the inverse-gamma prior.}

\item{init_sigma}{(Optional) A positive number; the starting value of the sigma parameter.}

\item{num_steps}{Integer; number of MCMC steps.}
}
\value{
A list of posterior samples of the parameters; each parameter is either a
a vector of length `num_steps`, or a matrix of dimension `num_steps` x d, where
d is the dimension of the parameter.
}
\description{
Model inference for normal regression model with multivariate normal
prior for the mean and inverse gamma for the variance.
}
\examples{
\dontrun{
n <- 100
p <- 10
data0 <- gaussian_data(n, p, intercept = T)
res <- gaussian_Gibbs(
  X = data0$X, y = data0$y,
  b_0 = rnorm(p+1), B_0 = pdmatrix(p+1)$Sigma,  # add one for the intercept
  alpha_0 = 13, delta_0 = 8,
)
}
}
