% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VAR_forecast.R
\name{predict_VAR}
\alias{predict_VAR}
\title{Predict future values given a VAR model}
\usage{
predict_VAR(n_step_ahead, data0, b_0, B)
}
\arguments{
\item{n_step_ahead}{integer; number of future forecast to generate.}

\item{data0}{Matrix; each row is one observation, each column is one measurement / predictor.}

\item{b_0}{numeric vector; the intercept.}

\item{B}{list of numeric matrices; the regression coefficients.}
}
\description{
Predict future values given a VAR model
}
