% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VAR.R
\name{VAR_Gibbs}
\alias{VAR_Gibbs}
\title{Model inference for Bayesian Vector-AutoRegressive (VAR) model with
normal-inverse-Wishart priors.}
\usage{
VAR_Gibbs(data0, p, b_0, V, v_0, S_0, init_Sigma, num_steps = 10000,
  burn_ins = 1000)
}
\arguments{
\item{data0}{A numeric matrix; a multivariate time series. The number
of columns is equal to the number of time points, the number of rows is
equal to the dimension of the observation. One can get such data from
the first component of the output from 'simulate_VAR'.}

\item{p}{integer; lag of the time series model.}

\item{b_0}{A numeric vector; the mean for the multivariate normal prior.}

\item{V}{A numeric matrix; the covariance for the multivariate normal prior.}

\item{v_0}{scalar; degree of freedom for the inverse-Wishart prior.}

\item{S_0}{matrix; the scale matrix for the inverse-Wishart prior.}

\item{init_Sigma}{(Optional) matrix; the starting value of the noise covariance.}

\item{num_steps}{integer; number of MCMC steps.}

\item{burn_ins}{integer; number of burn-ins.}
}
\description{
Model inference for Bayesian Vector-AutoRegressive (VAR) model with
normal-inverse-Wishart priors.
}
