% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VAR_forecast.R
\name{VAR_posterior_forecast}
\alias{VAR_posterior_forecast}
\title{Compute the posterior statistics of the parameters}
\usage{
VAR_posterior_forecast(n_step_ahead, data0, model0)
}
\arguments{
\item{n_step_ahead}{integer; number of future forecast to generate.}

\item{data0}{numeric matrix; the multivariate time series data.
The matrix must have at least as many columns as the lag, and the first
column should be the state at the first time point.}

\item{model0}{Output from 'VAR_Gibbs'.}
}
\description{
Compute the posterior statistics of the parameters
}
