% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wishart_inference.R
\name{wishart_Gibbs}
\alias{wishart_Gibbs}
\title{Model inference for Wishart regression model with multivariate normal
priors for the means and inverse gamma for the variances.}
\usage{
wishart_Gibbs(Xy, y, b_0, B_0, Xs, s, g_0, G_0, v_0, R_0, init_gamma,
  init_sigma, num_steps = 10000, burn_ins = 1000)
}
\arguments{
\item{Xy}{A numeric matrix; the first set of covariates.}

\item{y}{A numeric vector; the response variable for the first set of covariates.}

\item{b_0}{A numeric vector; the mean for the first multivariate normal prior.}

\item{B_0}{A numeric matrix; the covariance for the first multivariate normal prior.}

\item{Xs}{A numeric matrix; the second set of covariates.}

\item{s}{A numeric vector; the response variable for the second set of covariates.}

\item{g_0}{A numeric vector; the mean for the second multivariate normal prior.}

\item{G_0}{A numeric matrix; the covariance for the second multivariate normal prior.}

\item{v_0}{A positive number; the degrees of freedom of the Wishart prior.}

\item{R_0}{A 2x2 symmetric matrix; the scale matrix of the Wishart prior.}

\item{init_gamma}{(Optional) A numeric vector; the starting value of the gamma parameter.}

\item{init_sigma}{(Optional) A 2x2 symmetric matrix; the starting value of the sigma parameter.}

\item{num_steps}{Integer; number of MCMC steps.}

\item{burn_ins}{Integer; number of burn-ins.}
}
\description{
Model inference for Wishart regression model with multivariate normal
priors for the means and inverse gamma for the variances.
}
\examples{
\dontrun{
n <- 1000
p <- 2
k <- 3
data0 <- wishart_data(n, p, k,
                      sigma = matrix(c(1, 0.2, 0.2, 1), 2, 2),
                      intercept_1 = TRUE, intercept_2 = TRUE)
res <- wishart_Gibbs(Xy = data0$Xy, y = data0$y,
                     b_0 = rnorm(p + 2), B_0 = pdmatrix(p + 2)$Sigma,
                     Xs <- data0$Xs, s = data0$s,
                     g_0 = rnorm(k + 1), G_0 = pdmatrix(k + 1)$Sigma,
                     v_0 = 5, R_0 = pdmatrix(2)$Sigma)
}
}
